Struts2 dynamic autocomplete using Jquery and oracle
Hi Friends hope everyone knows autocomplete in jquery. It may be
difficult for us to implement autocomplete in STRUTS2. Believe me  it is
as simple as static autocomplete in jquery. I will show you how to
implement dynamic (data is coming from oracle) autocomplete in struts2
with the help of JQUERY